Move tests old (pydvl.value) vs. new (pydvl.valuation) to old testing module to make new testing module independent of pydvl.value
