+++ tests/valuation/methods/test_classwise_shapley.py
@@ -5,12 +5,8 @@ import logging
 import numpy as np
 import pytest
 from numpy.typing import NDArray
-from sklearn import datasets
-from sklearn.linear_model import LogisticRegression
 from typing_extensions import Self
 
-from pydvl.utils.dataset import Dataset as OldDataset
-from pydvl.utils.utility import Utility as OldUtility
 from pydvl.valuation.dataset import Dataset
 from pydvl.valuation.methods import ClasswiseShapleyValuation
 from pydvl.valuation.result import ValuationResult
@@ -18,17 +14,10 @@ from pydvl.valuation.samplers import (
     ClasswiseSampler,
     DeterministicPermutationSampler,
     DeterministicUniformSampler,
-    PermutationSampler,
-    UniformSampler,
 )
 from pydvl.valuation.scorers import ClasswiseSupervisedScorer
 from pydvl.valuation.stopping import MaxUpdates
 from pydvl.valuation.utility import ClasswiseModelUtility
-from pydvl.value import MaxChecks
-from pydvl.value.shapley.classwise import ClasswiseScorer as OldClasswiseScorer
-from pydvl.value.shapley.classwise import compute_classwise_shapley_values
-from pydvl.value.shapley.truncated import NoTruncation
-from tests.value import check_values
 
 from .. import check_values
 
@@ -173,49 +162,3 @@ def test_classwise_shapley(
     valuation.fit(train_dataset_manual_derivation)
     values = valuation.values()
     check_values(values, exact_solution, **check_kwargs)
-
-
-@pytest.mark.flaky(reruns=2)
-@pytest.mark.parametrize("n_samples", [500], ids=lambda x: "n_samples={}".format(x))
-def test_old_vs_new(
-    n_samples: int,
-    seed,
-):
-    model = LogisticRegression(random_state=seed)
-    old_data = OldDataset.from_sklearn(
-        datasets.load_iris(),
-        train_size=0.05,
-        random_state=seed,
-        stratify_by_target=True,
-    )
-    old_scorer = OldClasswiseScorer("accuracy", initial_label=0)
-    old_u = OldUtility(model=model, data=old_data, scorer=old_scorer)
-    old_values = compute_classwise_shapley_values(
-        old_u,
-        done=MaxChecks(n_samples),
-        truncation=NoTruncation(),
-        done_sample_complements=MaxChecks(1),
-        seed=seed,
-    )
-
-    new_train_data = Dataset(old_data.x_train, old_data.y_train)
-    new_test_data = Dataset(old_data.x_test, old_data.y_test)
-
-    in_class_sampler = PermutationSampler(seed=seed)
-    out_of_class_sampler = UniformSampler(seed=seed)
-    sampler = ClasswiseSampler(
-        in_class=in_class_sampler,
-        out_of_class=out_of_class_sampler,
-    )
-    new_u = ClasswiseModelUtility(
-        model,
-        ClasswiseSupervisedScorer("accuracy", new_test_data),
-        catch_errors=False,
-    )
-    valuation = ClasswiseShapleyValuation(
-        new_u,
-        sampler=sampler,
-        is_done=MaxUpdates(n_samples),
-    )
-    valuation.fit(new_train_data)
-    check_values(valuation.values(), old_values, atol=1e-1, rtol=1e-1)
